Support non-blocking i/o across platforms
This change introduces new tests for `O_NONBLOCK` and `SOCK_NONBLOCK` to
confirm that non-blocking i/o is now working on all supported platforms,
including Windows. For example, you can now say on Windows, MacOS, etc.:

    socket(AF_INET, SOCK_STREAM | SOCK_NONBLOCK, IPPROTO_TCP);

To create a non-blocking IPv4 TCP socket. Or you can enable non-blocking
i/o on an existing socket / pipe / etc. file descriptor by calling fcntl

    fcntl(fd, F_SETFL, fcntl(fd, F_GETFL) | O_NONBLOCK);

This functionality is polyfilled on older Linux kernels too, e.g. RHEL5.
Now that fcntl() support is much better the FIOCLEX / FIONCLEX polyfills
for ioctl() have been removed since they're ugly non-POSIX diameond APIs

This change fixes a weakness in kprintf() that was causing Windows trace
tools to frequently crash.

+++ libc/calls/getdomainname.c
@@ -76,6 +76,6 @@ int getdomainname(char *name, size_t len) {
   if (!rc && len && !strcmp(name, "(none)")) {
     name[0] = 0;
   }
-  STRACE("getdomainname([%#.*s], %'zu) → %d% m", len, name, len, rc);
+  STRACE("getdomainname([%#.*s], %'zu) → %d% m", (int)len, name, len, rc);
   return rc;
 }
